split up execution context module
branch-name: asset-execution-context
